Implemented \fsmilli (Apple extension)
git-svn-id: https://osxcart.svn.sourceforge.net/svnroot/osxcart/trunk@58 b2c7ee7d-4edc-4b10-941b-b19b4c2b722c

+++ osxcart/rtf-deserialize.h
@@ -14,7 +14,7 @@ typedef gpointer StateNewFunc(void);
 typedef gpointer StateCopyFunc(gpointer);
 typedef void StateFreeFunc(gpointer);
 
-#define POINTS_TO_PANGO(pts) (pts * PANGO_SCALE)
+#define POINTS_TO_PANGO(pts) ((gint)(pts * PANGO_SCALE))
 #define HALF_POINTS_TO_PANGO(halfpts) (halfpts * PANGO_SCALE / 2)
 #define TWIPS_TO_PANGO(twips) (twips * PANGO_SCALE / 20)
